requirements.txt
hydra-core==1.2.0
hydra-colorlog==1.2.0
omegaconf==2.2.3
matplotlib~=3.5.2
numpy~=1.23.0
xmlschema~=1.11.3
laspy~=2.3.0
pandas~=1.4.3
polyscope~=1.3.1
tqdm~=4.64.0
scipy~=1.8.1
lxml~=4.9.1
trimesh~=3.16.0
